[DDW-814] Rename DiscreetToggle component
