+++ source/renderer/app/features/discreet-mode/index.js
@@ -7,7 +7,7 @@ export {
   DiscreetValue,
   DiscreetTokenWalletAmount,
   DiscreetWalletAmount,
-  DiscreetToggle,
+  DiscreetModeToggle,
   DiscreetToggleTopBar,
   withDiscreetMode,
 } from './ui';
