--- source/renderer/app/features/discreet-mode/ui/DiscreetModeFeatureInject.js
@@ -1,17 +1,0 @@
-// @flow
-import type { Node } from 'react';
-import { observer } from 'mobx-react';
-import { useDiscreetModeFeature } from '../context';
-import { DiscreetMode } from '../feature';
-
-type Props = {
-  children: (discreetModeFeature: DiscreetMode) => Node,
-};
-
-const DiscreetModeFeatureInject = ({ children }: Props) => {
-  const discreetModeFeature = useDiscreetModeFeature();
-
-  return children(discreetModeFeature);
-};
-
-export default observer(DiscreetModeFeatureInject);
