+++ source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetModeToggle.scss
