+++ source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetModeToggle.story.js
@@ -3,17 +3,20 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import { DiscreetToggleButton } from './DiscreetToggle';
+import { DiscreetModeToggleComponent } from './DiscreetModeToggle';
 
-storiesOf('Discreet Mode|Discreet Toggle', module)
+storiesOf('Discreet Mode|Discreet Mode Toggle', module)
   .addDecorator(withKnobs)
   .add('Main', () => (
     <div style={{ padding: 20 }}>
       <div style={{ marginBottom: 20 }}>
-        <DiscreetToggleButton onToggle={action('onChange')} isDiscreetMode />
+        <DiscreetModeToggleComponent
+          onToggle={action('onChange')}
+          isDiscreetMode
+        />
       </div>
       <div>
-        <DiscreetToggleButton
+        <DiscreetModeToggleComponent
           onToggle={action('onChange')}
           isDiscreetMode={false}
         />
