+++ source/renderer/app/features/discreet-mode/ui/index.js
@@ -3,6 +3,6 @@
 export { default as DiscreetValue } from './DiscreetValue';
 export { default as DiscreetTokenWalletAmount } from './DiscreetTokenWalletAmount';
 export { default as DiscreetWalletAmount } from './DiscreetWalletAmount';
-export { default as DiscreetToggle } from './discreet-toggle/DiscreetToggle';
+export { DiscreetModeToggle } from './discreet-toggle/DiscreetModeToggle';
 export { default as DiscreetToggleTopBar } from './discreet-toggle-top-bar/DiscreetToggleTopBar';
 export { withDiscreetMode } from './withDiscreetMode';
