+++ storybook/stories/index.js
@@ -44,5 +44,5 @@ import './common/Widgets.stories';
 import './common/ItemsDropdown.stories';
 
 // Discreet Mode
-import '../../source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetToggle.story';
+import '../../source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetModeToggle.story';
 import '../../source/renderer/app/features/discreet-mode/ui/DiscreetValue.story';
